strawk: fix remaining test failures
- printf %a/%A on unix
- *printf invalid free
- cat() wrong results and corruption in alloc from invalid write
- sub/gsub erroneous logic with now static buffer
- always use int for adjbuf bufsize

+++ strawk/tran.c
@@ -588,7 +588,7 @@ Cell *catstr(Cell *a, Cell *b) /* concatenate a and b */
 	char *sb = getsval(b);
 	size_t l = strlen(sa) + strlen(sb) + 1;
 	static char *buf;
-	static size_t bufsz;
+	static int bufsz;
 
 	l++;	// add room for ' '
 	if(buf == NULL){
